verify jwt in post
