+++ src/backend/controller/controller.go
@@ -12,6 +12,8 @@ import (
 func SetupRoutes(router *httprouter.Router, dbConn *sql.DB) {
 	router.POST("/register", applyDbMiddleware(dbConn, service.HandleRegister))
 	router.POST("/login", applyDbMiddleware(dbConn, service.HandleLogin))
+	router.POST("/todo", applyDbMiddleware(dbConn, service.HandleCreateSimpleTodo))
+	//router.GET("/list/:userId", 	applyDbMiddleware())
 }
 
 func applyDbMiddleware(dbConn *sql.DB, next httprouter.Handle) httprouter.Handle {
